fix(tup-cms): pill size too large (on cms)
It is too small on portal, relative to table size. But table size on portal (matches core-styles.base) is smaller than cms (overrides core-styles.base). So, to avoid too-tiny pill text on portal, I let it be the same size as table. When portal text size is increased, the inconsistency can be fixed.