feat: Core-CMS#v4.24 & Core-Style#v2.39 (#499)
* feat: TACC/Core-CMS#v4.24.0 & Core-CMS#911

https://github.com/TACC/Core-CMS/releases/tag/v4.24.0

https://github.com/TACC/Core-CMS/pull/911

* fix: v4.24.1 not released, use branch docker image

* chore: save navbar tweak snippet before #499

* chore: save navbar tweak snippet after #499

Header CSS changed, so this snippet must change.

* fix: navbar tweak snippet typos

* chore: core-cms v4.24.1

Previous image was for PR/brnach that has been merged and released.

* style: add missing space

* Update Dockerfile

* build: use cms version tag

* deps: core-cms v4.24.2

Previous update (v4.24.1) only had dep update, but not jquery fix. I was scattered yesterday.

* chore: core-styles v2.39.0

* Revert "chore: core-styles v2.39.0"

This reverts commit d2b7d313abed4c681f4ad173f6ad7a86f586ed2a.

That commit (or the merge after it) triggered GitHub Actions that failed. I don't think it was related, but gonna revert and re-run, just to be sure.

* chore: rename a snippet

(Renamed on dev and prod.)

* fix: nav width overflow on CMS page, logged out

- logged in is fine, cuz no "Log in" text
- Portal is fine, because different font (Roboto) (a bug, btw)

* chore: core-styles v2.39.0

---------

Co-authored-by: Jake Rosenberg <jrosenberg@tacc.utexas.edu>